+++ frontend/packages/client/src/components/modals/VoteConfirmed.js
